Разнесла по папкам, удалила старый файл
